use exp instead of expf

+++ ising2D_cl.cl
@@ -2,7 +2,7 @@
 //~ #include <iomanip>
 //~ #include <cstdlib>
 //~ #include <cstdio>
-#include <cmath>
+//~ #include <cmath>
 //~ #include <limits>
 //~ #include <vector>
 //~ #include <sys/time.h>
@@ -89,7 +89,7 @@ inline bool mucaUpdate(float rannum, int* energy, char* d_lattice, unsigned idx,
 
   // flip with propability W(E_new)/W(E_old)
   // weights are stored in texture memory for faster random access
-  if (rannum < expf(tex1Dfetch(t_log_weights, EBIN(*energy + dE, d_N)) - tex1Dfetch(t_log_weights, EBIN(*energy, d_N)))) {
+  if (rannum < exp(tex1Dfetch(t_log_weights, EBIN(*energy + dE, d_N)) - tex1Dfetch(t_log_weights, EBIN(*energy, d_N)))) {
     d_lattice[idx * d_NUM_WORKERS + WORKER] = -d_lattice[idx * d_NUM_WORKERS + WORKER];
     *energy += dE;
     return true;
